server_script/requirements.txt
absl-py==2.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
cython-bbox==0.1.5
filelock==3.15.4
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lap==0.4.0
lazy_loader==0.4
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
minio==7.2.7
motmetrics==1.4.0
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pygame==2.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
xmltodict==0.13.0
zipp==3.19.2
